fix: modal-handlers.js module added, eslint errors used before declaration fixed
